feat: #120 KingFisher를 Rx로 래핑한 함수 Utility로 이동
