+++ WalWal/GlobalState/Project.swift
@@ -33,7 +33,7 @@ let project = Project(
       dependencies: [
         .ThirdParty.RxSwift,
         .ThirdParty.RxCocoa,
-        .ThirdParty.Kingfisher
+        .Utility
       ]
     ),
   ]
